+++ docs/NEWS.md
@@ -1,5 +1,3 @@
-# Changelog
-
 ## packer 0.0.7.9000
 
 - Ensure existing R files are not overwritten is already existing.
@@ -8,6 +6,7 @@ callable from outside the bundle.
 - Allow `@import` statements for CSS loader
 - Changed improved docs
 - Add `modules` and `import` arguments to CSS loader function.
+- Add `use_loader_style` to easily allow CSS modules.
 
 ## packer 0.0.6